Scripting MVP testable

+++ src/main/java/org/correomqtt/business/fileprovider/ScriptingProvider.java
@@ -1,11 +1,16 @@
 package org.correomqtt.business.fileprovider;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.apache.commons.io.FileUtils;
+import org.correomqtt.business.scripting.ExecutionDTO;
+import org.correomqtt.business.scripting.ScriptExecutionError;
 import org.correomqtt.business.scripting.ScriptFileDTO;
+import org.correomqtt.business.scripting.ScriptingBackend;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -15,6 +20,8 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.stream.Stream;
+
+import static org.apache.commons.io.FilenameUtils.removeExtension;
 
 public class ScriptingProvider extends BaseUserFileProvider {
 
@@ -34,7 +41,12 @@ public class ScriptingProvider extends BaseUserFileProvider {
     }
 
     private String getScriptFolder() {
-        return getTargetDirectoryPath() + File.separator + SCRIPT_FOLDER;
+        return getFromCache(SCRIPT_FOLDER);
+    }
+
+    public String getScriptLogFolder(String filename) {
+        return getFromCache(SCRIPT_LOG_FOLDER_NAME + File.separator + filename);
+
     }
 
     public List<ScriptFileDTO> getScripts() throws IOException {
@@ -47,8 +59,10 @@ public class ScriptingProvider extends BaseUserFileProvider {
             Files.createDirectory(scriptPath);
         }
 
+        List<ScriptFileDTO> scripts;
+
         try (Stream<Path> pathStream = Files.walk(scriptPath)) {
-            return pathStream
+            scripts = pathStream
                     .filter(f -> Files.isRegularFile(f) && f.getFileName().toString().endsWith(".js"))
                     .map(f -> {
                         LOGGER.info("Found script \"{}\"", f.toAbsolutePath());
@@ -59,6 +73,35 @@ public class ScriptingProvider extends BaseUserFileProvider {
                     })
                     .toList();
         }
+
+        scripts.forEach(script -> {
+            ObjectMapper om = new ObjectMapper();
+            om.registerModule(new JavaTimeModule());
+            String scriptExecutionFolder = getScriptExecutionsDirectory(script.getName());
+            try (Stream<Path> pathStream = Files.walk(new File(scriptExecutionFolder).toPath())) {
+                pathStream
+                        .filter(f -> Files.isRegularFile(f) && f.getFileName().toString().endsWith(".json"))
+                        .filter(f -> ScriptingBackend.getExecutionDTO(removeExtension(f.getFileName().toString())) == null)
+                        .forEach(f -> {
+                            try {
+                                ExecutionDTO dto = om.readValue(f.toFile(), ExecutionDTO.class);
+                                dto.setScriptFile(script);
+                                if (dto.getExecutionTime() == null) {
+                                    dto.setExecutionTime(0L);
+                                    dto.setError(new ScriptExecutionError(ScriptExecutionError.Type.HOST, "Unfinished"));
+                                    dto.setCancelled(true);
+                                }
+                                ScriptingBackend.putExecutionDTO(dto.getExecutionId(), dto);
+                            } catch (IOException e) {
+                                LOGGER.error("Unable to load execution file for {}", f.getFileName(), e);
+                            }
+                        });
+            } catch (IOException e) {
+                LOGGER.error("Unable to load execution files for folder {}", scriptExecutionFolder);
+            }
+        });
+
+        return scripts;
     }
 
     public Path createScript(String filename, String content) throws IOException {
@@ -80,9 +123,29 @@ public class ScriptingProvider extends BaseUserFileProvider {
         String absoluteNewFilename = getScriptFolder() + File.separator + newName;
         File oldFile = new File(absoluteOldFilename);
         File newFile = new File(absoluteNewFilename);
+        File oldExecutionDir = new File(getScriptExecutionsDirectory(oldName, false));
+        File newExecutionDir = new File(getScriptExecutionsDirectory(newName, false));
+        File oldLogDir = new File(getScriptLogDirectory(oldName, false));
+        File newLogDir = new File(getScriptLogDirectory(newName, false));
+        if (newFile.exists()) {
+            throw new FileAlreadyExistsException(absoluteNewFilename);
+        }
+        if (newExecutionDir.exists()) {
+            throw new FileAlreadyExistsException(newExecutionDir.getAbsolutePath());
+        }
+        if (newLogDir.exists()) {
+            throw new FileAlreadyExistsException(newLogDir.getAbsolutePath());
+        }
         if (!oldFile.renameTo(newFile)) {
             throw new FileAlreadyExistsException(absoluteNewFilename);
         }
+        if (oldExecutionDir.exists() && !oldExecutionDir.renameTo(newExecutionDir)) {
+            throw new FileAlreadyExistsException(newExecutionDir.getAbsolutePath());
+        }
+        if (oldLogDir.exists() && !oldLogDir.renameTo(newLogDir)) {
+            throw new FileAlreadyExistsException(newLogDir.getAbsolutePath());
+        }
+
         return newFile.toPath();
     }
 
@@ -99,6 +162,12 @@ public class ScriptingProvider extends BaseUserFileProvider {
         String absoluteFilename = getScriptFolder() + File.separator + filename;
         File file = new File(absoluteFilename);
         Files.delete(file.toPath());
+        deleteExecutions(filename);
+    }
+
+    public void deleteExecutions(String filename) throws IOException {
+        FileUtils.deleteDirectory(new File(getScriptLogFolder(filename)));
+        FileUtils.deleteDirectory(new File(getScriptExecutionsDirectory(filename)));
     }
 
     public void saveScript(ScriptFileDTO scriptFileDTO, String content) throws IOException {
@@ -106,5 +175,30 @@ public class ScriptingProvider extends BaseUserFileProvider {
         try (FileWriter fileWriter = new FileWriter(scriptFileDTO.getPath().toFile())) {
             fileWriter.write(content);
         }
+    }
+
+    public String getSingleScriptLogPath(String filename, String executionId) {
+        return getScriptLogFolder(filename) + File.separator + executionId + ".log";
+    }
+
+    public String loadLog(String filename, String executionId) throws IOException {
+        String logFolder = getSingleScriptLogPath(filename, executionId);
+        StringBuilder codeBuilder = new StringBuilder();
+        File file = new File(logFolder);
+
+        if (!file.exists()) {
+            return null;
+        }
+
+        try (Stream<String> lines = Files.lines(file.toPath(), StandardCharsets.UTF_8)) {
+            lines.forEach(s -> codeBuilder.append(s).append("\n"));
+            return codeBuilder.toString().strip();
+        }
+    }
+
+    public void saveExecution(ExecutionDTO dto) throws IOException {
+        ObjectMapper om = new ObjectMapper();
+        om.registerModule(new JavaTimeModule());
+        om.writeValue(new File(getScriptExecutionsDirectory(dto.getScriptFile().getName()) + File.separator + dto.getExecutionId() + ".json"), dto);
     }
 }
